requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
imageio-ffmpeg==0.4.8
kombu==5.2.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pycodestyle==2.10.0
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
redis==4.5.4
rfc3986==1.5.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.6
